pyproject.toml
[tool.poetry]
name = "minijob-client"
version = "0.1.0"
description = "A client library for accessing FastAPI"

authors = []

readme = "README.md"
packages = [
    {include = "minijob_client"},
]
include = ["CHANGELOG.md", "minijob_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.6"
httpx = "^0.15.0"
attrs = "^20.1.0"
python-dateutil = "^2.8.1"

[tool.black]
line-length = 120
target_version = ['py36', 'py37', 'py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"